requirements.txt
django-autocomplete-light==3.7.0.dev0
